requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.4
python-multipart>=0.0.9
sqlalchemy>=2.0.30
pydantic>=2.7.0
pydantic-settings>=2.2.1
openai>=1.80.0
rank-bm25>=0.2.2
pypdf>=5.0.0
python-docx>=1.1.2
pandas>=2.2.0
openpyxl>=3.1.0
